renames package pipeline => pipelines
