+++ src/dimcat/__init__.py
@@ -15,7 +15,7 @@ from .base import DimcatConfig, get_class, get_schema
 from .data.dataset.base import Dataset
 from .steps.extractors.base import FeatureExtractor
 from .steps.loaders.base import PackageLoader
-from .steps.pipeline.base import Pipeline
+from .steps.pipelines.base import Pipeline
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
